pyproject.toml
[project]
name = "image-generation-model"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.28.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.1",
]
